Refactor dev and cybersecurity pages into MVC
