+++ views/templates/header.php
@@ -42,8 +42,8 @@ global $api;
                     <li><a href="index.php" <?php echo !isset($_GET['route']) || $_GET['route'] == 'home' ? 'class="active"' : ''; ?>>Accueil</a></li>
                     <li><a href="index.php?route=forums" <?php echo isset($_GET['route']) && $_GET['route'] == 'forums' ? 'class="active"' : ''; ?>>Forums</a></li>
                     <li><a href="index.php?route=articles" <?php echo isset($_GET['route']) && $_GET['route'] == 'articles' ? 'class="active"' : ''; ?>>Articles</a></li>
-                    <li><a href="cyber-securite.php" <?php echo basename($_SERVER['PHP_SELF']) == 'cyber-securite.php' ? 'class="active"' : ''; ?>>Cyber-Sécurité</a></li>
-                    <li><a href="dev.php" <?php echo basename($_SERVER['PHP_SELF']) == 'dev.php' ? 'class="active"' : ''; ?>>Développement</a></li>
+                    <li><a href="index.php?route=cyber-securite" <?php echo isset($_GET['route']) && $_GET['route'] == 'cyber-securite' ? 'class="active"' : ''; ?>>Cyber-Sécurité</a></li>
+                    <li><a href="index.php?route=dev" <?php echo isset($_GET['route']) && $_GET['route'] == 'dev' ? 'class="active"' : ''; ?>>Développement</a></li>
                     <li><a href="membres.php" <?php echo basename($_SERVER['PHP_SELF']) == 'membres.php' ? 'class="active"' : ''; ?>>Membres</a></li>
                 </ul>
             </nav>
